@@ -75,8 +75,8 @@
         <version>3.8.1</version>
         <configuration>
           <encoding>UTF-8</encoding>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>11</source>
+          <target>11</target>
           <compilerArgument>-Werror</compilerArgument>
         </configuration>
       </plugin>
