@@ -260,13 +256,17 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-          <compilerVersion>1.8</compilerVersion>
+          <source>11</source>
+          <target>11</target>
+          <compilerVersion>11</compilerVersion>
           <debug>true</debug>   <!-- Whether to include debugging information.   -->
           <encoding>UTF-8</encoding> <!-- The -encoding argument for the Java compiler. -->
-          <compilerArgument>-proc:none
-          </compilerArgument> <!-- disable automatic annotation processing accoring to: http://docs.jboss.org/hibernate/jpamodelgen/1.0/reference/en-US/html_single/#d0e261 -->
+          <compilerArgs>
+            <!-- disable automatic annotation processing according to: http://docs.jboss.org/hibernate/jpamodelgen/1.0/reference/en-US/html_single/#d0e261 -->
+            <arg>-proc:none</arg>
+            <arg>--add-exports</arg>
+            <arg>java.base/sun.net.ftp=ALL-UNNAMED</arg>
+          </compilerArgs>
         </configuration>
       </plugin>
 
