@@ -48,7 +48,9 @@ import org.springframework.util.StringUtils;
 import jeeves.component.ProfileManager;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 /**
  * @author ETj (etj at geo-solutions.it)
@@ -153,9 +155,6 @@ public class ShibbolethUserUtils {
                 user = (User) authProvider.loadUserByUsername(username);
 
                 if (config.isUpdateGroup()) {
-                    // First we remove all previous groups
-                    userGroupRepository.deleteAll(UserGroupSpecs.hasUserId(user.getId()));
-
                     // Now we add the groups
                     assignGroups(groupRepository, userGroupRepository, roleGroups,
                             roleGroupSeparator, user);
@@ -222,6 +221,9 @@ public class ShibbolethUserUtils {
 
     private void assignGroups(GroupRepository groupRepository, UserGroupRepository userGroupRepository,
                               String[] role_groups, String separator, User user) {
+
+        Set<UserGroup> userGroups =  new HashSet<>();
+
         // Assign groups
         int i = 0;
 
@@ -258,14 +260,16 @@ public class ShibbolethUserUtils {
                     ug.setGroup(g);
                     ug.setUser(user);
                     ug.setProfile(Profile.Editor);
-                    userGroupRepository.save(ug);
+                    userGroups.add(ug);
                 }
             } else {
                 // Failback if no profile
                 usergroup.setProfile(Profile.Guest);
             }
-            userGroupRepository.save(usergroup);
+            userGroups.add(usergroup);
         }
+
+        userGroupRepository.updateUserGroups(user.getId(), userGroups);
     }
 
     private void assignProfile(String[] role_groups, String roleGroupSeparator, User user) {
