@@ -22,6 +22,7 @@
 //==============================================================================
 package org.fao.geonet.doi.client;
 
+import com.google.common.base.Function;
 import com.google.common.io.CharStreams;
 import org.apache.commons.httpclient.HttpStatus;
 import org.apache.commons.io.IOUtils;
@@ -31,11 +32,18 @@ import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.BasicCredentialsProvider;
+import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.message.BasicHeader;
+import org.fao.geonet.ApplicationContextHolder;
+import org.fao.geonet.kernel.setting.SettingManager;
+import org.fao.geonet.lib.Lib;
 import org.fao.geonet.utils.GeonetHttpRequestFactory;
 import org.fao.geonet.utils.Log;
 import org.springframework.http.client.ClientHttpResponse;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.io.InputStreamReader;
 
 import static org.fao.geonet.doi.client.DoiSettings.LOGGER_NAME;
@@ -62,6 +70,7 @@ public class BaseDoiClient {
 
             postMethod = new HttpPost(url);
 
+
             ((HttpUriRequest) postMethod).addHeader( new BasicHeader("Content-Type",  contentType + ";charset=UTF-8") );
             Log.debug(LOGGER_NAME, "   -- Request body: " + body);
 
@@ -72,9 +81,8 @@ public class BaseDoiClient {
 
             postMethod.setEntity(requestEntity);
 
-            httpResponse = requestFactory.execute(
-                postMethod,
-                new UsernamePasswordCredentials(username, password), AuthScope.ANY);
+            httpResponse = executeRequest(postMethod);
+
             int status = httpResponse.getRawStatusCode();
 
             Log.debug(LOGGER_NAME, "   -- Request status code: " + status);
@@ -92,7 +100,7 @@ public class BaseDoiClient {
                     successMessage, url));
             }
         } catch (Exception ex) {
-            Log.error(LOGGER_NAME, "   -- Error (exception): " + ex.getMessage());
+            Log.error(LOGGER_NAME, "   -- Error (exception): " + ex.getMessage(), ex);
             throw new DoiClientException(ex.getMessage());
 
         } finally {
@@ -116,9 +124,8 @@ public class BaseDoiClient {
 
             getMethod = new HttpGet(url);
 
+            httpResponse = executeRequest(getMethod);
 
-            httpResponse = requestFactory.execute(getMethod,
-                new UsernamePasswordCredentials(username, password), AuthScope.ANY);
             int status = httpResponse.getRawStatusCode();
 
             Log.debug(LOGGER_NAME, "   -- Request status code: " + status);
@@ -137,7 +144,7 @@ public class BaseDoiClient {
             }
 
         } catch (Exception ex) {
-            Log.error(LOGGER_NAME, "   -- Error (exception): " + ex.getMessage());
+            Log.error(LOGGER_NAME, "   -- Error (exception): " + ex.getMessage(), ex);
             throw new DoiClientException(ex.getMessage());
 
         } finally {
@@ -147,5 +154,27 @@ public class BaseDoiClient {
             // Release the connection.
             IOUtils.closeQuietly(httpResponse);
         }
+    }
+
+
+    protected ClientHttpResponse executeRequest(HttpUriRequest method) throws Exception {
+        final String requestHost = method.getURI().getHost();
+
+        final Function<HttpClientBuilder, Void> requestConfiguration = new Function<HttpClientBuilder, Void>() {
+            @Nullable
+            @Override
+            public Void apply(@Nonnull HttpClientBuilder input) {
+                final BasicCredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+                credentialsProvider.setCredentials(AuthScope.ANY, new UsernamePasswordCredentials(username, password));
+                input.setDefaultCredentialsProvider(credentialsProvider);
+
+                Lib.net.setupProxy(ApplicationContextHolder.get().getBean(SettingManager.class), input, requestHost);
+                input.useSystemProperties();
+
+                return null;
+            }
+        };
+
+        return requestFactory.execute(method, requestConfiguration);
     }
 }
