@@ -166,7 +166,7 @@ public class EsHTTPProxy {
                 context,
                 context.getBean(IMetadataUtils.class)
                     .findOneByUuid(doc.get("_id").asText()),
-                relatedTypes, 0, 100);
+                relatedTypes, 0, 1000);
         } catch (Exception e) {
             LOGGER.warn("Failed to load related types for {}. Error is: {}",
                 getSourceString(doc, Geonet.IndexFieldNames.UUID),
