@@ -45,14 +45,6 @@
       this.queue = queue;
       this.errors = errors;
 
-      var clear = function (map) {
-        var type = (map && map.get && map.get("type")) || "viewer";
-        if (queue[type]) {
-          queue[type].queue.length = 0;
-          queue[type].errors.length = 0;
-        }
-      };
-
       var getMapType = function (map) {
         var type = (map && map.get && map.get("type")) || "viewer";
         if (queue[type] === undefined) {
@@ -132,6 +124,21 @@
           style: style
         };
         return getLayerIndex(queue[getMapType(map)], layer) >= 0;
+      };
+
+      /**
+       *
+       * @param {ol.Map} map
+       */
+      this.clear = function (map) {
+        var type = (map && map.get && map.get("type")) || "viewer";
+        if (queue[type]) {
+          queue[type] = [];
+        }
+
+        if (errors[type]) {
+          errors[type] = [];
+        }
       };
     }
   ]);
