@@ -99,13 +99,13 @@
       this.getFormatterPath = function (defaultFormatter) {
         var tokens = $location.path().split("/");
         if (tokens.length > 2 && tokens[3] === "formatters") {
-          return "../api/records/" + $location.url().split(/^metadraf|metadata\//)[1];
+          return "../api/records/" + $location.url().split(/^\/(metadraf|metadata)\//)[2];
         } else if (tokens.length > 2 && tokens[3] === "main") {
           return undefined; // Angular view
         } else if (defaultFormatter) {
           return (
             "../api/records/" +
-            $location.url().split(/^metadraf|metadata\//)[1] +
+            $location.url().split(/^\/(metadraf|metadata)\//)[2] +
             defaultFormatter
           );
         } else {
