@@ -661,6 +661,7 @@
           scope.orderById = attrs.orderById || "false";
           scope.max = gnThesaurusService.DEFAULT_NUMBER_OF_RESULTS;
           scope.fauxMultilingual = scope.fauxMultilingual === "true"; //default false
+          scope.showHintsOnFocus = attrs.showHintsOnFocus === "true"; // displays all the values on focus, default shows only the selected value
 
           // Configuration only required when using the directive in template fields.
           //
@@ -910,7 +911,8 @@
               .typeahead(
                 {
                   minLength: 0,
-                  highlight: true
+                  highlight: true,
+                  showHintsOnFocus: scope.showHintsOnFocus
                 },
                 {
                   name: "keyword",
